+++ libs/synthefy/src/synthefy/__init__.py
@@ -17,7 +17,7 @@ from synthefy.data_models import (
 )
 from synthefy.nori_client import SynthefyNoriClient
 
-__version__ = "7.1.2"
+__version__ = "7.1.3"
 
 __all__ = [
     "DEFAULT_LARGE_CONTEXT_SEED",